simulate all commits
ae-simulate functions does close immitation of the real conversion
process by manipulating with file names and RCS revision numbers. It
checks if repository is correct and all commits are applicable and it
finds start of every branch found in project. Return value is
association list of branch names and commit names where it should
start to apply cleanly.